ProjectManager: If project file does not exist, remove from recent files list.

+++ src/Project/ProjectManager.cpp
@@ -172,11 +172,16 @@ ProjectManager::removeProject( const QString& fileName )
     // Remove all occurence of fileName
     m_recentsProjects.removeAll( fileName );
 
+    QSettings s;
+    s.setValue( "RecentsProjects", m_recentsProjects );
+
+    QFile   projectFile( fileName );
+    if ( !projectFile.exists() )
+        return;
+
     if( !QFile::remove( fileName ) )
         QMessageBox::warning( NULL, tr( "Can't delete project file" ),
                           tr( "Can't delete this project file: %1\nPlease delete it manually." ).arg( fileName ) );
-    QSettings s;
-    s.setValue( "RecentsProjects", m_recentsProjects );
 }
 
 void
